virtio-mmio: firecracker specific handling
Kata used a disk for the VM rootfs in the case of Firecracker.
(Instead of initrd or virtual-nvdimm).

The Kata code today does not handle this case properly.

For now as Firecracker is the only Hypervisor in Kata that
uses virtio-mmio directly offset the drive index to comprehend
this.

Longer term we should track if the rootfs is setup as a block
device explicitly.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>